@@ -51,7 +51,7 @@
     "jest-environment-jsdom": "27.4.6",
     "lodash": "4.17.21",
     "namor": "2.0.4",
-    "next": "13.1.6",
+    "next": "13.4.0",
     "parse-link-header": "2.0.0",
     "postcss": "8.4.14",
     "react": "18.2.0",
