@@ -1,6 +1,6 @@
 import { MenuItem } from '@heathmont/moon-core-tw';
 import NextLink from 'next/link';
-import { useRouter } from 'next/router';
+import { usePathname } from 'next/navigation';
 
 type Props = {
   href: string;
@@ -9,7 +9,7 @@ type Props = {
 };
 
 const Link: React.FC<Props> = ({ href, children, onClick }) => {
-  const { pathname } = useRouter();
+  const pathname = usePathname();
   const isActive = href === pathname;
   return (
     <NextLink href={href} passHref legacyBehavior>
