@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import {
   ThemeProvider,
@@ -146,7 +147,7 @@ export const DocsThemeContext = React.createContext<DocsThemeContextFields>({
 
 export const useDocsTheme = () => React.useContext(DocsThemeContext);
 
-export const DocsThemeProvider: React.FC<{ children?: React.ReactElement }> = ({
+export const DocsThemeProvider: React.FC<{ children?: React.ReactNode }> = ({
   children,
 }) => {
   // TODO Save mode in local storage
